Let merge_subPR deal with submodules not on main
merge_subPR only works if the PR of the submodule has
already been merged. However, it may be that the local
repository is not on the main branch resulting from
that merge.
With this change we now take care of this case and
update the local repository to the respective merge
commit, for closing the PR in the super repository.