Fix coco in variants
The generation of the coco.set file during configuration, does not
work properly with variant builds, as the coco.set file is created
in the default build directory.
We now generate the coco.set file during the build phase and also
add a dependency and the necessity to create the coco.set file
before any coco task is being run.
Furthermore, search for the coco settings file in parent directories
of nested wscript setups, to allow the use of one in the top most
directory.